feat: Move UI validations to shared-ui module
Refs: XRDDEV-3072
