+++ src/shared-ui/src/utils/helpers.ts
@@ -45,3 +45,14 @@ function toClipboard(val: string): void {
 export const helper = {
   toClipboard,
 };
+
+// Checks if the given WSDL URL valid
+export function isValidWsdlURL(str: string): boolean {
+  const pattern = new RegExp('(^(https?):///?)[-a-zA-Z0-9]');
+  return pattern.test(str);
+}
+
+// Checks if the given REST URL is valid
+export function isValidRestURL(str: string): boolean {
+  return isValidWsdlURL(str);
+}
